+++ src/main/kotlin/cn/yiiguxing/plugin/translate/trans/openai/OpenAiTranslator.kt
@@ -4,7 +4,7 @@ import cn.yiiguxing.plugin.translate.message
 import cn.yiiguxing.plugin.translate.service.CacheService
 import cn.yiiguxing.plugin.translate.trans.*
 import cn.yiiguxing.plugin.translate.trans.openai.chat.ChatRole
-import cn.yiiguxing.plugin.translate.trans.openai.chat.chatCompletionRequest
+import cn.yiiguxing.plugin.translate.trans.openai.chat.chatMessages
 import cn.yiiguxing.plugin.translate.trans.openai.exception.OpenAIStatusException
 import cn.yiiguxing.plugin.translate.ui.settings.TranslationEngine.OPEN_AI
 import cn.yiiguxing.plugin.translate.util.md5
@@ -13,7 +13,7 @@ import com.intellij.openapi.components.service
 import org.jsoup.nodes.Document
 import javax.swing.Icon
 
-object OpenAITranslator : AbstractTranslator(), DocumentationTranslator {
+object OpenAiTranslator : AbstractTranslator(), DocumentationTranslator {
 
     override val id: String = OPEN_AI.id
     override val name: String = OPEN_AI.translatorName
@@ -22,14 +22,14 @@ object OpenAITranslator : AbstractTranslator(), DocumentationTranslator {
     override val contentLengthLimit: Int = OPEN_AI.contentLengthLimit
     override val primaryLanguage: Lang get() = OPEN_AI.primaryLanguage
     override val supportedSourceLanguages: List<Lang> =
-        OpenAILanguages.languages.toMutableList().apply { add(0, Lang.AUTO) }
-    override val supportedTargetLanguages: List<Lang> = OpenAILanguages.languages
+        OpenAiLanguages.languages.toMutableList().apply { add(0, Lang.AUTO) }
+    override val supportedTargetLanguages: List<Lang> = OpenAiLanguages.languages
 
-    private val openAIModel: OpenAIModel get() = service<OpenAISettings>().model
+    private val settings: OpenAiSettings get() = service<OpenAiSettings>()
 
 
     override fun checkConfiguration(force: Boolean): Boolean {
-        if (force || !OpenAICredential.isApiKeySet) {
+        if (force || !OpenAiCredentials.manager(settings.provider).isCredentialSet) {
             return OPEN_AI.showConfigurationDialog()
         }
 
@@ -58,16 +58,15 @@ object OpenAITranslator : AbstractTranslator(), DocumentationTranslator {
         targetLang: Lang,
         isFofDocumentation: Boolean
     ): String {
-        val model = openAIModel
         val cacheService = service<CacheService>()
-        val cacheKey = getCacheKey(model, text, srcLang, targetLang)
+        val cacheKey = getCacheKey(text, srcLang, targetLang)
         val cache = cacheService.getDiskCache(cacheKey)
         if (!cache.isNullOrEmpty()) {
             return cache
         }
 
-        val request = getChatCompletionRequest(model, text, srcLang, targetLang, isFofDocumentation)
-        val chatCompletion = OpenAI.chatCompletion(request)
+        val messages = getChatCompletionMessages(text, srcLang, targetLang, isFofDocumentation)
+        val chatCompletion = OpenAiService.get(settings.getOptions()).chatCompletion(messages)
         var result = chatCompletion.choices.first().message!!.content
         if (!isFofDocumentation && result.length > 1 && result.first() == '"' && result.last() == '"') {
             result = result.substring(1, result.lastIndex)
@@ -77,38 +76,39 @@ object OpenAITranslator : AbstractTranslator(), DocumentationTranslator {
         return result
     }
 
-    private fun getChatCompletionRequest(
-        openAIModel: OpenAIModel,
+    private fun getChatCompletionMessages(
         text: String,
         srcLang: Lang,
         targetLang: Lang,
         isFofDocumentation: Boolean = false
-    ) =
-        chatCompletionRequest {
-            model = openAIModel.value
-            messages {
-                message {
-                    role = ChatRole.SYSTEM
-                    content = "You are a translation engine that can " + if (isFofDocumentation) {
-                        "translate HTML document."
-                    } else {
-                        "only translate text and cannot interpret it."
-                    }
-                }
-                message {
-                    role = ChatRole.USER
-                    content =
-                        "Translate ${if (srcLang == Lang.AUTO) "" else "from ${srcLang.openAILanguage} "}to ${targetLang.openAILanguage}."
-                }
-                message {
-                    role = ChatRole.USER
-                    content = if (isFofDocumentation) text else """"$text""""
-                }
+    ) = chatMessages {
+        message {
+            role = ChatRole.SYSTEM
+            content = "You are a translation engine that can " + if (isFofDocumentation) {
+                "translate HTML document."
+            } else {
+                "only translate text and cannot interpret it."
             }
         }
+        message {
+            role = ChatRole.USER
+            content =
+                "Translate ${if (srcLang == Lang.AUTO) "" else "from ${srcLang.openAiLanguage} "}to ${targetLang.openAiLanguage}."
+        }
+        message {
+            role = ChatRole.USER
+            content = if (isFofDocumentation) text else """"$text""""
+        }
+    }
 
-    private fun getCacheKey(model: OpenAIModel, text: String, srcLang: Lang, targetLang: Lang): String {
-        return "$id$model$text$srcLang$targetLang".md5()
+    private fun getCacheKey(text: String, srcLang: Lang, targetLang: Lang): String {
+        val provider = settings.provider
+        val model = when (val options = settings.getOptions(provider)) {
+            is OpenAiService.OpenAIOptions -> options.model.value
+            is OpenAiService.AzureOptions -> options.deployment ?: ""
+        }
+
+        return "$id$provider$model$text$srcLang$targetLang".md5()
     }
 
     override fun createErrorInfo(throwable: Throwable): ErrorInfo? {