+++ src/main/kotlin/cn/yiiguxing/plugin/translate/ui/TranslationBalloon.kt
@@ -4,6 +4,7 @@ import cn.yiiguxing.plugin.translate.*
 import cn.yiiguxing.plugin.translate.service.TranslationUIManager
 import cn.yiiguxing.plugin.translate.trans.Lang
 import cn.yiiguxing.plugin.translate.trans.Translation
+import cn.yiiguxing.plugin.translate.tts.TTSEngine
 import cn.yiiguxing.plugin.translate.ui.balloon.BalloonImpl
 import cn.yiiguxing.plugin.translate.ui.balloon.BalloonPopupBuilder
 import cn.yiiguxing.plugin.translate.ui.icon.Spinner
@@ -229,6 +230,10 @@ class TranslationBalloon(
     }
 
     override fun onTranslatorChanged(settings: Settings, translationEngine: TranslationEngine) {
+        hide()
+    }
+
+    override fun onTTSEngineChanged(settings: Settings, ttsEngine: TTSEngine) {
         hide()
     }