+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Translation Plugin Changelog
 
 ## [Unreleased]
+- Add OpenAI TTS engine.
+- OpenAI Translator now supports Azure OpenAI.
+- OpenAI Translator now offers API endpoint configuration.
+- The model list of OpenAI Translator has been updated.
 
 ## [3.5.8] (2024/04/10)